Moved subdiv code to HolidayUtil
